+++ awkernel_drivers/src/lib.rs
@@ -3,7 +3,7 @@
 extern crate alloc;
 
 pub mod hal;
-pub mod interrupt_controler;
+pub mod interrupt_controller;
 pub mod net;
 pub mod uart;
